chore: move AirdropJob types to types folder
