+++ src/back/models/AirdropJob.ts
@@ -1,6 +1,8 @@
 import { Model } from 'decentraland-gatsby/dist/entities/Database/model'
 import { SQL, table } from 'decentraland-gatsby/dist/entities/Database/utils'
 
+import { AirdropJobStatus } from '../types/AirdropJob'
+
 export type AirdropJobAttributes = {
   id: string
   badge_spec: string
@@ -10,14 +12,6 @@ export type AirdropJobAttributes = {
   created_at: Date
   updated_at: Date
 }
-
-export enum AirdropJobStatus {
-  PENDING = 'pending',
-  FINISHED = 'finished',
-  FAILED = 'failed',
-}
-
-export type AirdropOutcome = Pick<AirdropJobAttributes, 'status' | 'error'>
 
 export default class AirdropJobModel extends Model<AirdropJobAttributes> {
   static tableName = 'airdrop_jobs'
