+++ src/back/types/AirdropJob.ts
@@ -1,0 +1,9 @@
+import { AirdropJobAttributes } from '../models/AirdropJob'
+
+export enum AirdropJobStatus {
+  PENDING = 'pending',
+  FINISHED = 'finished',
+  FAILED = 'failed',
+}
+
+export type AirdropOutcome = Pick<AirdropJobAttributes, 'status' | 'error'>
