+++ src/entities/Badges/types.ts
@@ -43,7 +43,6 @@ export enum ActionStatus {
 
 export type RevokeOrReinstateResult = { status: ActionStatus; address: string; badgeId: string; error?: string }
 export type BadgeCreationResult = { status: ActionStatus; badgeCid?: string; error?: string; badgeTitle?: string }
-
 export type GasConfig = { gasPrice: ethers.BigNumber; gasLimit: ethers.BigNumber }
 export const GAS_MULTIPLIER = 2
